third_party/WebKit/LayoutTests/fast/html/imports/import-nested-dup-2.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
var marks = [];
function start()
{
    var url = document.currentScript.ownerDocument.URL;
    var path = url.match(/[^\/]+$/)
    marks.push("START:" + path);
}

function end()
{
    var url = document.currentScript.ownerDocument.URL;
    var path = url.match(/[^\/]+$/)
    marks.push("END  :" + path);
}

start();
</script>
<link id="harness" rel="import" href="resources/nest-dup-grand-parent.html">
<link id="harness" rel="import" href="resources/nest-dup-2-parent.html">
</head>
<body>
<script>
end();
description("Test that a subimport should be loaded even if is linked as a top-level import.");
var expected = ["START:import-nested-dup-2.html",
                "START:nest-dup-grand-parent.html",
                "START:nest-dup-parent.html",
                "START:nest-dup.html",
                "START:nest-dup-child.html",
                "END  :nest-dup-child.html",
                "END  :nest-dup.html",
                "END  :nest-dup-parent.html",
                "END  :nest-dup-grand-parent.html",
                "START:nest-dup-2-parent.html",
                "START:nest-dup-2.html",
                "END  :nest-dup-2.html",
                "END  :nest-dup-2-parent.html",
                "END  :import-nested-dup-2.html"]
for (var i = 0; i < expected.length; ++i)
    shouldBe("marks[i]", "expected[i]");
shouldBe("marks.length", "expected.length");
</script>
</body>
</html>
