third_party/WebKit/LayoutTests/fast/html/imports/import-nested-dup-2-expected.txt

Test that a subimport should be loaded even if is linked as a top-level import.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks.length is expected.length
PASS successfullyParsed is true

TEST COMPLETE
